General improvements to tests within project (#1)
This change includes the addition of unit tests for the UserService to assist with interviewees. In addition it also include a change to the file name of the existing test so it represents the fact it is a integration test.

This change also includes rectifying a comment on the seed data to point to the correct record that is overdue.